+++ session/src/init/templates/kv/models.ts
@@ -1,5 +1,5 @@
 import { type KvValue, model } from "@olli/kvdex";
-import type { SessionDoc } from "@innovatedev/fresh-session/kvdex-store";
+import { sessionModel } from "@innovatedev/fresh-session/kvdex-store";
 
 // Define your Session data model
 // You can extend this with additional properties if needed
@@ -11,5 +11,6 @@ export type User = {
   // {{USER_FIELDS}}
 } & KvValue;
 
-export const SessionModel = model<SessionDoc<SessionData>>();
+// export const SessionModel = sessionModel(type({ theme: "'light' | 'dark'" }));
+export const SessionModel = sessionModel<SessionData>();
 export const UserModel = model<User>();